All tests passed

+++ server/services/BirdsService.js
@@ -7,7 +7,7 @@ class BirdsService {
     if (bird.creatorId !== userId) {
       throw new BadRequest('You are not the creator you can not delete this bird')
     }
-    const birdDeleted = await dbContext.Bird.findOneAndDelete({ _id: id, creatorId: userId })
+    const birdDeleted = await dbContext.Bird.findOneAndDelete({ _id: id, creatorId: userId }).populate('creator')
     return birdDeleted
   }
 
@@ -16,7 +16,7 @@ class BirdsService {
     if (bird.creatorId !== userId) {
       throw new BadRequest('You are not the creator you can not edit this.')
     }
-    const editedBird = await dbContext.Bird.findOneAndUpdate({ _id: id, creatorId: userId }, body)
+    const editedBird = await dbContext.Bird.findOneAndUpdate({ _id: id, creatorId: userId }, body).populate('creator')
     return editedBird
   }
 
@@ -26,7 +26,7 @@ class BirdsService {
   }
 
   async findById(id) {
-    const bird = await dbContext.Bird.findById(id)
+    const bird = await dbContext.Bird.findById(id).populate('creator')
     if (!bird) {
       throw new BadRequest('Invalid Id --> Bird does not exist')
     }
@@ -34,7 +34,7 @@ class BirdsService {
   }
 
   async find(query = {}) {
-    const birds = await dbContext.Bird.find(query)
+    const birds = await dbContext.Bird.find(query).populate('creator')
     if (!birds) {
       throw new BadRequest('Invalid Query Request')
     }
